Add a base class for events with a WindowModel source.
For #662, it makes sense to have a standard way to
get the source from events we may wish to filter.

This introduces a SourcedEvent interface, and adapts classes
to use it.
